Fix wrong class definition on user change Split controller base away from controller test
